Rename from aio to nimbus
Hopefully this will be enough to park the name for awhile.

Nimbus is intended to be a codename, in the vein of Greyskull or
Burgundy, which does not convey any particular meaning outside of
serving as an identifier. Most of the names this project has been called
over the last two weeks had some kind of meaning associated with them,
which has caused problems and spurred additional renames. Ideally nimbus
will let us just park it here for awhile, either until it is meaningful
as a project name itself or until we decide on a long-term name.