Greatly simplify nimbus face
The four options originally built in were not generally useful and so
served primarily only to overcomplicate the use case. This commit
removes the options, replacing the environment logic with enough to
allow the usually Puppet --environment and --environmentpath options to
work if specified by the user.

If any of the advanced or persistant config functionality is ever
actually needed, it can be easily added back OR a purpose-built face
created that extends the nimbus base for the use case.