Fix source=puppet:// file parameters
Previously when a file resource specified the source parameter with a
puppet:// protocol value, because the rest indirector was selected for
resolving content and metadata. This didn't work because there was no
Puppet server.

This commit fixes the issue by specify another Puppet setting to
override, :default_file_terminus = :file_server.