Correct how hocon gem is required
See https://tickets.puppetlabs.com/browse/HC-47 for details.
Essentially, if we want to use Hocon::ConfigFactory, we need to `require
"hocon/config_factor"`, not just `require "hocon"`. The library used to
work if we did the latter, but updates being released in the 1.0.0 hocon
gem will make that not work anymore.